bot/requirements.txt
aiogram==3.17.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
Requests==2.32.3
spotify_music_dl==0.0.3
yt_dlp==2024.12.23
